Test 3 9 14

+++ app/DTO/Investments/fixedTermDTO.php
@@ -8,7 +8,7 @@ class fixedTermDTO
      * Create a new class instance.
      */
     public function __construct(
-        public readonly int $monto,
+        public readonly float $monto,
         public readonly int $plazo,
 
     ) {}
